Update and rename localization/english/culture/spark_cultural_traditions_l_english.yml to localization/english/culture/traditions/spark_cultural_traditions_l_english.yml
Bringing this file in line with Vanilla practices too by splitting the Traditions from the Tradition Effects. Though instead of just leaving everything in one big pile like Vanilla, clear delineations have been retained for ease of updating/expanding in the future.

Also relocating to the traditions subfolder for consistency, as that's where it's Vanilla equivalent is located.