Update and rename atlantis_nicknames_l_english.yml to elf_nicknames_l_english.yml
Localisations seem to vary with their prefixes, and it's starting to bother me. Also a ton of duplications and other stuff. Like how this localisation exists in the "names" subfolder (which is not where this localisation file should be).

As far as the prefixes go, I'm going to start changing all the prefixes to elf. I thought about using elfdestiny, but that ended up feeling too long. This effort will also change the localisations prefixed with "spark" as that's an element of the elven faith, rather than the mod itself.