Update elf_nicknames_l_english.yml
I've added back Vana and Tulkas, along with Ryumplstilkin, as I wasn't exactly sure where better would fit for these keys.

And while Ryumpl had his own accompanying description key, Vana and Tulkas did not, so I drummed up a couple for them. Not that I have any idea if they'd even be useful.

Also, to avoid conflicts with the hidden name key for Ryumpl, I changed this key to have the "entity_" prefix. Not sure if that'll mess with anything. I'll go dig for references to him in the files before I initiate my PR.