Update elf_nicknames_l_english.yml
This is why we research, always. Ryumpl is a nickname for some random Irish Elf! Lame, it would've been cooler to have the proper Ryumpl causing mischief, but oh well. I've moved him out of the Shroud of Mysteries here and back up into the general ED nickname pool to be drawn from for the pre-generated characters.